refactor implementation
